Update connection.js
This update returns with check complete ,but .env variable returns as promise object ,so on performing node server.js it outputs with mongooseerror: , as the mongoose.connect() expects the string but, object got passed. The check is successful bcoz in test I had passed the url explicitly and stubbed console.log and mongoose.connect to check if the correct statement gets printed in output. Since I had passed the url in string format to mongoose.connect ,the test got passesd.